www/requirements.txt
Flask
google-cloud-tasks
